requirements.txt
kafka-python==2.0.2
boto3==1.34.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-dotenv==1.0.0
pydantic==2.5.0
schedule==1.2.1
faker==22.0.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
